Alternative implementation of the OpticC category.
Using a type family to make the various flavours of lenses. This spreads
out the composition operators to the individual implementations, which
is sensible. It also introduces a class that can be used for cloning
lenses in a generalised manner.

The main drawback is extra unwrapping/rewrapping work in the public
module. Well, at least that gives a bit more rationale for having
dedicated wrapper-operators around every thing there, instead of
just re-exporting...